+++ services/src/api/handlers/spatial_references.rs
@@ -270,7 +270,7 @@ mod tests {
     use crate::contexts::Session;
     use crate::ge_context;
     use crate::pro::contexts::PostgresContext;
-    use crate::pro::users::UserAuth;
+    use crate::users::UserAuth;
     use crate::util::tests::send_test_request;
     use actix_web;
     use actix_web::http::header;
