+++ services/src/api/handlers/upload.rs
@@ -202,7 +202,7 @@ mod tests {
     use crate::contexts::Session;
     use crate::ge_context;
     use crate::pro::contexts::PostgresContext;
-    use crate::pro::users::UserAuth;
+    use crate::users::UserAuth;
     use crate::util::tests::{send_test_request, SetMultipartBody, TestDataUploads};
     use actix_web::{http::header, test};
     use actix_web_httpauth::headers::authorization::Bearer;
