+++ services/src/error.rs
@@ -432,7 +432,7 @@ pub enum Error {
 
     #[snafu(context(false), display("OidcError: {}", source))]
     Oidc {
-        source: crate::pro::users::OidcError,
+        source: crate::users::OidcError,
     },
 
     #[snafu(display(
