+++ services/src/permissions/mod.rs
@@ -2,8 +2,8 @@ use crate::error::{self, Error, Result};
 use crate::identifier;
 use crate::layers::listing::LayerCollectionId;
 use crate::machine_learning::MlModelId;
-use crate::pro::users::UserId;
 use crate::projects::ProjectId;
+use crate::users::UserId;
 use async_trait::async_trait;
 use geoengine_datatypes::dataset::{DatasetId, LayerId};
 use postgres_types::{FromSql, ToSql};
