+++ services/src/users/session.rs
@@ -3,8 +3,8 @@ use crate::contexts::{ApplicationContext, Session, SessionId};
 use crate::error;
 use crate::permissions::{Role, RoleId};
 use crate::pro::contexts::PostgresContext;
-use crate::pro::users::UserId;
 use crate::projects::{ProjectId, STRectangle};
+use crate::users::UserId;
 use crate::util::Identifier;
 use actix_http::Payload;
 use actix_web::{web, FromRequest, HttpRequest};
