+++ services/src/users/userdb.rs
@@ -2,10 +2,10 @@ use crate::api::handlers::users::UsageSummaryGranularity;
 use crate::contexts::SessionId;
 use crate::error::Result;
 use crate::permissions::{RoleDescription, RoleId};
-use crate::pro::users::oidc::{OidcTokens, UserClaims};
-use crate::pro::users::{UserCredentials, UserId, UserRegistration, UserSession};
 use crate::projects::{ProjectId, STRectangle};
 use crate::quota::{ComputationQuota, DataUsage, DataUsageSummary, OperatorQuota};
+use crate::users::oidc::{OidcTokens, UserClaims};
+use crate::users::{UserCredentials, UserId, UserRegistration, UserSession};
 use async_trait::async_trait;
 use geoengine_datatypes::primitives::DateTime;
 use geoengine_operators::meta::quota::ComputationUnit;
