+++ services/src/util/openapi_examples.rs
@@ -1,7 +1,7 @@
 use crate::api::model::responses::ErrorResponse;
 use crate::contexts::SessionId;
 use crate::pro::contexts::PostgresContext;
-use crate::pro::users::UserAuth;
+use crate::users::UserAuth;
 use actix_web::dev::ServiceResponse;
 use actix_web::http::{header, Method};
 use actix_web::test::TestRequest;
